requirements.txt
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
docker==7.1.0
exceptiongroup==1.3.1
filetype==1.2.0
google-ai-generativelanguage==0.11.0
google-api-core==2.30.2
google-auth==2.49.1
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-core==0.3.83
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
SQLAlchemy==2.0.49
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
